input: touchscreen: siw: update bus driver control
The type of chip_pdata for each device is changed to 'const'
to preserve its own definitions.
Instead, the new structure of siw_touch_chip_data is defined
to transfer chip_pdata to driver core layer.
In the bus setup state, an additional memory is allocated
and used as dupliacted data of the 'const' chip_pdata.
So, the core doesn't directly access the original 'const' chip_pdata.