input: touchscreen: siw: reorganize lcd mode management
The 'mode_allowed' is a smart way to check supported mode range,
however, the real code structure is not smart
because the mode bit which is used as main key instaed of mode index
is very confusing point.

This patch removes this confusion
by restoring the mode index to main control key.

The related commit history
- 96dee5ac89095f19166fc70883abe6dbd5e150f9
  : add condition check for lcd mode control
- 0f7db211bcbeeb83a26fba0acc4694a6310a2aab
  : bug-fix lcd_mode (bit field)
- 6e17b7697ec9ca6a83ac27f951981a93db4a6234
  : add sysfs attribute