input: touchscreen: siw: bug-fix for reg address collision
When it comes to LG4895,
reg->prd_m1_m2_raw_offset is reassigned to 0x2D6 from 0x287.
This 0x2D6 is the default value of reg->ext_watch_lut for watch layer.
This means that reg->ext_watch_lut isn't available in LG4895 and
it requires an exception control to avoid address collision.