input: touchscreen: siw: hal sysfs - new feature for reg. debugging history
The sysfs reg_ctrl is useful for register test or tracing.
But, the echo command has the limit that can't show the kernel log directly in adb shell.

This patch add new structure to save reg. debugging history for better convenience,
which shows max 8 debug result of 'echo ... > reg_ctrl'

.../siw_touch_input # cat reg_ctrl
[Test History]
 rd: reg[0x0201] = 0x0F5201E7
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000
 __: reg[0x0000] = 0x00000000

* add reg_list by seperating it from the former _show_reg_ctrl